# swagger.yml
swagger: "2.0"
info:
  description: "This is a sample server DigiControl server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters."
  version: "1.0.0"
  title: "Swagger DigiControl.com"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "apiteam@swagger.io"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:4000"
basePath: "/api"
...
schemes:
- "http"
paths:
...
  /users:
    get:
      tags:
      - "/api/users"
      summary: "Get all users"
      description: "Filter is not required"
      operationId: "getAll"
      produces:
      - "application/xml"
      - "application/json"
      responses:
        default:
          description: "successful operation"
    post:
      tags:
      - "/api/users"
      summary: "Create user"
      description: "This can only be done by the logged in user."
      operationId: "createUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Created user object"
        required: true
        schema:
          $ref: "#/definitions/UserCreate"
      responses:
        default:
          description: "Succesfully Created User"
    put:
      tags:
      - "/api/users"
      summary: "Update user"
      description: "This can only be done by the logged in user."
      operationId: "updateUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Updated user object"
        required: true
        schema:
          $ref: "#/definitions/UserCreate"
      responses:
        "200":
          description: "Succesfully Updated User"
        "400":
          description: "Invalid user supplied"
        "404":
          description: "Error occured while Finding the User"
    delete:
      tags:
      - "/api/users"
      summary: "Delete user"
      description: "This can only be done by the logged in user."
      operationId: "deleteUser"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "id"
        in: "query"
        description: "Id user to delete"
        required: true
        type: "string"
      responses:
        "200":
          description: "Succesfully Deleted"
        "404":
          description: "User Could not be deleted"
...
definitions:
...
  UserCreate:
    type: "object"
    properties:
      name:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
    xml:
      name: "UserCreate"
...